limit wrap size to max int

+++ jaccall.runtime/src/main/java/org/freedesktop/jaccall/JNI.java
@@ -121,7 +121,7 @@ public final class JNI {
      * JNI ->
      */
     public static native ByteBuffer wrap(long address,
-                                         @Nonnegative long size);
+                                         @Nonnegative int size);
 
     public static native long unwrap(@Nonnull Buffer byteBuffer);
